+++ mcumgr-toolkit/tests/common/mod.rs
@@ -1,4 +1,5 @@
 #![allow(unused)]
+#![allow(clippy::collapsible_if)]
 
 use std::{
     collections::VecDeque,